You are given three arrays containing the id, name, and image of users.
Objectives
Create a function with a name createUserObjects that takes three arrays as a parameter and returns an array of objects representing the user.
Use the map function to create objects from the arrays.
The mapping should be done in such a way that for the first object, the first element of all arrays should be used, and so on.