You are required to build a user authentication system for a web application Using the IIFE. Create an empty array of objects  with name users  to store registered users. The IIFE should return two functions. registerUser: It should take two parameters username and password and add it to the users array. Call the checkCredentials  function  to determine if the user is already present in the users array.
If the user is already registered then return  otherwise push the user object to the Users array and return .
checkCredentials: It should take two arguments, username and password, and check whether the user with the given credentials is present in the user's array or not. It should return false if the user is not present otherwise true.