You have been hired to develop a tracking system for a logistics company. The system should allow the company to keep track of the location and status of each shipment, as well as assign resources to each shipment as needed.  Create a constructor function called Shipment with properties such as
*Note: All the properties should be assigned with the values at the time of the object creation id: a string representing the shipment ID

location: string representing the location of shipment

destination: string representing destination for the shipment

status : string representing the status of the shipment

resources : an array containing all the resources that have been assigned to the shipment

Using the prototype of the Shipment constructor function, add the following methods
updateStatusandResources :
should take status and resources as arguments and update the status and resource property of the object.

assignResources : This method should be able to push the arguments passed to the resource array.

(Note: the number of arguments passed by the user is not defined. It can be any number of arguments)
Create an object called TrackingSystem using an object literal that has the following:
shipments(property): representing an array of shipment object

updateStatus(method) : It should take two parameters, id and status, and should be able to change the status of the object using the id of the object.
(Hint: You can use various array methods to iterate over the array)

viewShipment(method) : It should take id as a parameter and should display the details of that shipment like id, status, resources, location, and destination.
** use the concept of destructuring in viewShipment() method to display all the properties.